+++ openeogeotrellis/config/config.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import abc
 import os
+from pathlib import Path
 from typing import List, Optional, Union
 
 import attrs
@@ -254,3 +255,18 @@ class GpsBackendConfig(OpenEoBackendConfig):
     # calrissian_image: Optional[str] = "ghcr.io/duke-gcb/calrissian/calrissian:0.17.1"
     calrissian_image: Optional[str] = "registry.stag.warsaw.openeo.dataspace.copernicus.eu/rand/calrissian:latest"
     calrissian_bucket: Optional[str] = "calrissian"
+
+    """
+    Inject S3 profiles and tokens that allow S3 access scoped to the Job execution.      
+    """
+    provide_s3_profiles_and_tokens: bool = False
+
+    """
+    Config file to generate identity information from this Openeo Instance
+    """
+    openeo_idp_details_file: Path = Path("/opt/.idp/idp_details.json")
+
+    """
+    Directory where config specific to job execution will be mounted for batch jobs.
+    """
+    batch_job_config_dir: Path = Path("/opt/job_config")