+++ CHANGELOG.md
@@ -14,6 +14,8 @@ without compromising stable operations.
 
 ## Unreleased
 
+- Initial support for S3 profiles and tokens during batch job execution ([#969](https://github.com/Open-EO/openeo-geopyspark-driver/pull/969))
+
 
 ## 0.56.0